+++ ovp_projects/models/apply.py
@@ -21,21 +21,44 @@ class Apply(models.Model):
   email = models.CharField(_('email'), max_length=200, blank=True, null=True)
   phone = models.CharField(_('phone'), max_length=30, blank=True, null=True)
 
+  def __init__(self, *args, **kwargs):
+    super(Apply, self).__init__(*args, **kwargs)
+    self.__original_status = self.status
+    self.__original_canceled = self.canceled
+
   def mailing(self, async_mail=None):
     return emails.ApplyMail(self, async_mail)
 
   def save(self, *args, **kwargs):
-    if self.canceled:
-      self.canceled_date = timezone.now()
-      self.status = 'unapplied'
-      self.mailing().sendUnappliedToVolunteer({'apply': self})
-      self.mailing().sendUnappliedToOwner({'apply': self})
-    else:
-      self.canceled_date = None
-      self.status = 'applied'
+
+    if self.pk == None:
+      # Object being created
       self.mailing().sendAppliedToVolunteer({'apply': self})
       self.mailing().sendAppliedToOwner({'apply': self})
+    else:
+      # Object being updated
+      if self.__original_canceled != self.canceled:
+        # self.canceled was modified
+        if self.canceled == True:
+          self.status = "unapplied"
+        if self.canceled == False:
+          self.status = "applied"
 
+      # Status can be set without modifying .canceled directly
+      # Therefore we reset values checked on the previous ifs
+      if self.__original_status != self.status:
+        if self.status == "unapplied":
+          self.canceled = True
+          self.canceled_date = timezone.now()
+          self.mailing().sendUnappliedToVolunteer({'apply': self})
+          self.mailing().sendUnappliedToOwner({'apply': self})
+        else:
+          self.canceled = False
+          self.canceled_date = None
+
+    # Update original values
+    self.__original_status = self.status
+    self.__original_canceled = self.canceled
     return_data = super(Apply, self).save(*args, **kwargs)
 
     # Updating project applied_count